feat: added k8s_custom_resources table
