+++ plugins/source/k8s/resources/services/batch/cron_jobs.go
@@ -3,11 +3,12 @@ package batch
 import (
 	"context"
 
+	v1 "k8s.io/api/batch/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	"github.com/cloudquery/cloudquery/plugins/source/k8s/client"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
 	"github.com/cloudquery/plugin-sdk/v4/transformers"
-	v1 "k8s.io/api/batch/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func CronJobs() *schema.Table {
@@ -21,7 +22,7 @@ func CronJobs() *schema.Table {
 }
 
 func fetchCronJobs(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
-	cl := meta.(*client.Client).Client().BatchV1().CronJobs("")
+	cl := meta.(*client.Client).Client().BatchV1().CronJobs(metav1.NamespaceAll)
 
 	opts := metav1.ListOptions{}
 	for {
