+++ plugins/source/k8s/resources/services/custom_resources/custom_resource_definitions_test.go
@@ -1,16 +1,17 @@
-package crd
+package custom_resources
 
 import (
 	"testing"
+
+	resource "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	apiextensionsclientset "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/cloudquery/cloudquery/plugins/source/k8s/client"
 	"github.com/cloudquery/cloudquery/plugins/source/k8s/mocks/apiextensions"
 	mocks "github.com/cloudquery/cloudquery/plugins/source/k8s/mocks/apiextensions/v1"
 	"github.com/cloudquery/plugin-sdk/v4/faker"
 	"github.com/golang/mock/gomock"
-	resource "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	apiextensionsclientset "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func createCRDs(t *testing.T, ctrl *gomock.Controller) apiextensionsclientset.Interface {
