+++ plugins/source/k8s/resources/services/policy/pod_disruption_budgets.go
@@ -3,18 +3,19 @@ package policy
 import (
 	"context"
 
+	policy "k8s.io/api/policy/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
 	"github.com/cloudquery/cloudquery/plugins/source/k8s/client"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
 	"github.com/cloudquery/plugin-sdk/v4/transformers"
-	policy "k8s.io/api/policy/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func PodDisruptionBudgets() *schema.Table {
 	return &schema.Table{
 		Name:      "k8s_policy_pod_disruption_budgets",
 		Resolver:  fetchPodDisruptionBudgets,
-		Multiplex: client.ContextNamespaceMultiplex,
+		Multiplex: client.ContextMultiplex,
 		Transform: client.TransformWithStruct(&policy.PodDisruptionBudget{}, transformers.WithPrimaryKeys("UID")),
 		Columns:   schema.ColumnList{client.ContextColumn},
 	}
@@ -22,7 +23,7 @@ func PodDisruptionBudgets() *schema.Table {
 
 func fetchPodDisruptionBudgets(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
 	c := meta.(*client.Client)
-	cl := c.Client().PolicyV1().PodDisruptionBudgets(c.Namespace)
+	cl := c.Client().PolicyV1().PodDisruptionBudgets(metav1.NamespaceAll)
 
 	opts := metav1.ListOptions{}
 	for {
